ocr_service/requirements.txt
google-generativeai>=0.8.0
pillow>=10.4.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

